init training :sparkles:
