+++ vusic/separation/training/training.py
@@ -1,0 +1,9 @@
+import SeparationDataset from vusic.utils.separation_dataset
+
+def main():
+    print("hello");
+
+if __name__ == '__main__':
+    main()
+
+
